Make hooking optional
This allows for there to be a build (hopefully) devoid of hacks.
There is a regression somewhere in this commit, however, that seems to
stop the GlobalState's ModuleInitializer from running on launch (waiting
for a dependency??). It still runs when the tool window is opened, but
the whole point is that we want it to run early. We may need to move the
actual patching routine elsewhere if it allows us to somehow avoid this
caveat.