Fix regression with background loading
It seems like EmuHawk won't actually fully load the assembly unless
these OptionalApi properties are here. I'm not totally sure why at the
moment, but bisecting changes shows it's these statements that allow the
assembly to load early. May need to figure out why so that a more robust
solution can be introduced in the future.