Remove background execution hook
It was a cool idea, but ultimately BizHawk does *intentionally* try to
prevent external tools from running before they are opened for
security reasons, so the likelyhood of being able to consistently
subvert the module process to then only do further hacky things is
probably just asking for instability.

I also just don't really like the idea of having to bundle Harmony in
with the plugin, feels sort of gross somehow.