Use UpdateValues instead of GeneralUpdate
Prior to 2.10, external tools didn't receive GeneralUpdate at all.
We still have the caveat that prior to the following PR and commit,
external tools didn't receive updates unless the core was advancing.
This leaves one of the most common desired patterns not possible which
is pretty awkward, but I'll think if there's a worthwhile solution.

https://github.com/TASEmulators/BizHawk/issues/3626
https://github.com/TASEmulators/BizHawk/commit/5eb2cd8cb128928fc4fd9af00010a5d74c99adcd